refactor: re-structture the files position
