--- myapp/src/components/FormPage1.tsx
@@ -1,11 +1,0 @@
-import React from "react"
-import { QuestionsPage } from "@/components/QuestionsPage"
-import { GlobalStateProvider } from "@/context/GlobalStateContext"
-
-export const FormPage1 = () => {
-  return (
-    <GlobalStateProvider>
-      <QuestionsPage />
-    </GlobalStateProvider>
-  )
-}
