+++ myapp/src/components/Sections/DisplayDetails/DisplayInputs/DispayPeople.tsx
