+++ myapp/src/components/Sections/DisplayDetails/DisplayInputs/DisplayBudget.tsx
