+++ myapp/src/components/Sections/DisplayDetails/DisplayInputs/DisplayEvent.tsx
