+++ myapp/src/components/Sections/DisplayDetails/DisplayInputs/DisplayMessage.tsx
