+++ myapp/src/components/Sections/DisplayDetails/DisplayInputs/DisplayName.tsx
