+++ myapp/src/components/Sections/DisplayDetails/DisplayInputs/DisplayTools.tsx
