+++ myapp/src/components/Sections/DisplayDetails/DisplaySelectedValues.tsx
@@ -1,14 +1,14 @@
 import React from "react"
-import DisplayPeople from "@/components/DisplayDetails/DisplayInputs/DispayPeople"
-import { DisplayAddress } from "@/components/DisplayDetails/DisplayInputs/DisplayAddress"
-import { DisplayBudget } from "@/components/DisplayDetails/DisplayInputs/DisplayBudget"
-import { DisplayDate } from "@/components/DisplayDetails/DisplayInputs/DisplayDate"
-import DisplayEvent from "@/components/DisplayDetails/DisplayInputs/DisplayEvent"
-import { DisplayMessage } from "@/components/DisplayDetails/DisplayInputs/DisplayMessage"
-import { DsiplayName } from "@/components/DisplayDetails/DisplayInputs/DisplayName"
-import { DisplayPhone } from "@/components/DisplayDetails/DisplayInputs/DisplayPhone"
-import { DisplayTools } from "@/components/DisplayDetails/DisplayInputs/DisplayTools"
-import { DisplayVenue } from "@/components/DisplayDetails/DisplayInputs/DisplayVenue"
+import DisplayPeople from "@/components/Sections/DisplayDetails/DisplayInputs/DispayPeople"
+import { DisplayAddress } from "@/components/Sections/DisplayDetails/DisplayInputs/DisplayAddress"
+import { DisplayBudget } from "@/components/Sections/DisplayDetails/DisplayInputs/DisplayBudget"
+import { DisplayDate } from "@/components/Sections/DisplayDetails/DisplayInputs/DisplayDate"
+import DisplayEvent from "@/components/Sections/DisplayDetails/DisplayInputs/DisplayEvent"
+import { DisplayMessage } from "@/components/Sections/DisplayDetails/DisplayInputs/DisplayMessage"
+import { DsiplayName } from "@/components/Sections/DisplayDetails/DisplayInputs/DisplayName"
+import { DisplayPhone } from "@/components/Sections/DisplayDetails/DisplayInputs/DisplayPhone"
+import { DisplayTools } from "@/components/Sections/DisplayDetails/DisplayInputs/DisplayTools"
+import { DisplayVenue } from "@/components/Sections/DisplayDetails/DisplayInputs/DisplayVenue"
 
 type DisplaySelectedValuesProps = {
   currentStep: number
