+++ myapp/src/components/Sections/DisplayDetails/ShowPhotoOfTheEvent.tsx
