--- myapp/src/components/fetchFunctions/fetchFunctions.js
@@ -1,63 +1,0 @@
-// Fetch functions
-
-export const createNewCustomer = async () => {
-  try {
-    const res = await fetch("https://api.food-c.co.kr/api/new_customer", {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    })
-
-    if (!res.ok) {
-      throw new Error("Failed to fetch new customer.")
-    }
-
-    const data = await res.json()
-    return data.ticket_number
-  } catch (error) {
-    console.log("Error in createNewCustomer:", error)
-    throw error
-  }
-}
-
-export const saveFormData = async (ticket_number, formData) => {
-  try {
-    const res = await fetch(
-      `https://api.food-c.co.kr/api/process_data/${ticket_number}`,
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(formData),
-      }
-    )
-    if (!res.ok) {
-      throw new Error("Failed to save form data.")
-    }
-
-    const data = await res.json()
-  } catch (error) {
-    console.log("Error in saveFormData:", error)
-    throw error
-  }
-}
-
-export const generatePdf = async (ticket_number) => {
-  try {
-    const res = await fetch(
-      `https://api.food-c.co.kr/api/generatepdf/${ticket_number}`,
-      {
-        method: "GET",
-      }
-    )
-
-    if (!res.ok) {
-      throw new Error("Failed to generate PDF.")
-    }
-  } catch (error) {
-    console.log("Error in generatePdf:", error)
-    throw error
-  }
-}
